Code/CustomerAgent/requirements.txt
# Microsoft Agent Framework (GA core + orchestrations latest beta)
agent-framework-core==1.0.0
agent-framework-openai==1.0.0
agent-framework-orchestrations>=1.0.0b260402

# Azure OpenAI client for MAF
openai>=1.50.0
azure-identity>=1.17.0

# MCP client (used by MAF's MCPTool)
mcp>=1.5.0

# HTTP
httpx>=0.27.0

# Web server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Azure Key Vault (cert for SSO)
azure-keyvault-secrets>=4.8.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
plotly>=5.0.0

# Application Insights (OpenTelemetry — replaces deprecated opencensus)
azure-monitor-opentelemetry-exporter>=1.0.0b30
opentelemetry-sdk>=1.25.0

# Streamlit UI
streamlit>=1.35.0
PyJWT>=2.8.0
cryptography>=42.0.0
requests>=2.31.0
